♻️ refactor(webui): 仪表盘首页 routes/index 状态机下沉为领域 hook
候选 3：routes/index.tsx 2124→1330 行。多源只读数据 + 手写 TTL 缓存，保留缓存不转 Query。

- 新建 routes/home/hooks：useDashboardData（含 dashboardDataCache + 伪进度条）、useBotStatus（botStatusCache + 30s 轮询 + 可见性刷新）、useFeatureStatus、useLocalCacheMetrics、useReviewStats、useMaibotVersion（GitHub 版本 + 一言原生 fetch）、useQuickShortcuts（快捷菜单 + 插件快捷 + localStorage）
- home/types.ts 抽出共享类型
- 子组件（recharts 图表/ExpressionReviewer/RestartOverlay/统计卡片）零改动
- 特征化测试 3/3 全绿
